added dynamic show for sort and filter bar
