+++ frontend/src/components/Gallery/AppBar/SortBtn.vue
@@ -1,0 +1,25 @@
+<script setup lang="ts">
+import type { Events } from "@/types/emitter";
+import type { Emitter } from "mitt";
+import { ref, inject } from "vue";
+
+// Props
+const emitter = inject<Emitter<Events>>("emitter");
+const isShowSortBar = ref(false);
+function showSortBar() {
+  emitter?.emit('sortBarShow', null)
+  isShowSortBar.value = !isShowSortBar.value
+}
+</script>
+
+<template>
+  <v-btn
+    class="ml-0"
+    variant="text"
+    rounded="0"
+    icon="mdi-sort"
+    :color="isShowSortBar ? 'romm-accent-1' : ''"
+    title="Order gallery"
+    @click="showSortBar"
+  />
+</template>
